@@ -4,83 +4,83 @@
   "version": "2.0.0",
   "author": "Michelangelo Partipilo <michelangelo@thebrandingplug.com>",
   "dependencies": {
-    "autoprefixer": "^9.5.1",
+    "autoprefixer": "^9.7.6",
     "babel-plugin-lodash": "^3.3.4",
     "babel-polyfill": "^6.26.0",
-    "bootstrap": "^4.3.1",
+    "bootstrap": "^4.4.1",
     "classnames": "^2.2.6",
     "font-awesome": "^4.7.0",
-    "gatsby": "^2.3.32",
-    "gatsby-image": "^2.0.40",
-    "gatsby-plugin-lodash": "^3.0.5",
-    "gatsby-plugin-netlify-cache": "^1.1.0",
-    "gatsby-plugin-postcss": "^2.0.7",
-    "install": "^0.12.2",
-    "lodash": "^4.17.11",
+    "gatsby": "^2.20.26",
+    "gatsby-image": "^2.3.4",
+    "gatsby-plugin-lodash": "^3.2.3",
+    "gatsby-plugin-netlify-cache": "^1.2.0",
+    "gatsby-plugin-postcss": "^2.2.3",
+    "install": "^0.13.0",
+    "lodash": "^4.17.15",
     "moment": "^2.24.0",
-    "node-sass": "^4.12.0",
-    "npm": "^6.9.0",
+    "node-sass": "^4.13.1",
     "pixrem": "^5.0.0",
     "prop-types": "^15.7.2",
-    "react": "^16.8.6",
+    "react": "^16.13.1",
     "react-disqus-comments": "^1.4.0",
-    "react-dom": "^16.8.6",
-    "react-fontawesome": "^1.6.1",
-    "react-helmet": "^5.2.0",
-    "react-images": "^0.5.19",
-    "react-interval": "^2.0.2",
+    "react-dom": "^16.13.1",
+    "react-fontawesome": "^1.7.1",
+    "react-helmet": "^6.0.0",
+    "react-images": "^1.1.3",
+    "react-interval": "^2.1.1",
     "react-measure": "^2.3.0",
-    "react-moment": "^0.9.2",
-    "react-photo-gallery": "^7.0.2",
-    "react-popper": "^1.3.3",
-    "react-responsive": "^6.1.1",
+    "react-moment": "^0.9.7",
+    "react-photo-gallery": "^8.0.0",
+    "react-popper": "^2.2.2",
+    "react-responsive": "^8.0.3",
     "react-scrollable-anchor": "^0.6.1",
-    "react-share": "^2.4.0",
+    "react-share": "^4.1.0",
     "react-twitter-widgets": "^1.7.1",
-    "reactstrap": "^8.0.0"
+    "reactstrap": "^8.4.1",
+    "sharp": "^0.25.2"
   },
   "devDependencies": {
     "babel-plugin-array-includes": "^2.0.3",
     "babel-plugin-transform-object-assign": "^6.22.0",
     "cli-glob": "^0.1.0",
-    "eslint": "^5.16.0",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-config-prettier": "^4.2.0",
-    "eslint-config-standard": "^12.0.0",
-    "eslint-plugin-import": "^2.17.2",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-node": "^8.0.1",
-    "eslint-plugin-prettier": "^3.0.1",
-    "eslint-plugin-promise": "^4.1.1",
-    "eslint-plugin-react": "^7.12.4",
-    "eslint-plugin-standard": "^4.0.0",
-    "gatsby-plugin-catch-links": "^2.0.13",
-    "gatsby-plugin-feed": "^2.1.1",
-    "gatsby-plugin-google-analytics": "^2.0.18",
-    "gatsby-plugin-manifest": "^2.0.29",
-    "gatsby-plugin-nprogress": "^2.0.10",
-    "gatsby-plugin-offline": "^2.0.25",
-    "gatsby-plugin-react-helmet": "^3.0.12",
-    "gatsby-plugin-sass": "^2.0.11",
-    "gatsby-plugin-sharp": "^2.0.35",
-    "gatsby-plugin-sitemap": "^2.0.12",
-    "gatsby-plugin-twitter": "^2.0.13",
-    "gatsby-remark-autolink-headers": "^2.0.16",
-    "gatsby-remark-copy-linked-files": "^2.0.12",
-    "gatsby-remark-images": "^3.0.11",
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb": "^18.1.0",
+    "eslint-config-prettier": "^6.10.1",
+    "eslint-config-standard": "^14.1.1",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^3.1.3",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-react": "^7.19.0",
+    "eslint-plugin-standard": "^4.0.1",
+    "gatsby-plugin-catch-links": "^2.2.3",
+    "gatsby-plugin-feed": "^2.4.3",
+    "gatsby-plugin-google-analytics": "^2.2.4",
+    "gatsby-plugin-manifest": "^2.3.5",
+    "gatsby-plugin-nprogress": "^2.2.3",
+    "gatsby-plugin-offline": "^3.1.4",
+    "gatsby-plugin-react-helmet": "^3.2.4",
+    "gatsby-plugin-sass": "^2.2.3",
+    "gatsby-plugin-sharp": "^2.5.6",
+    "gatsby-plugin-sitemap": "^2.3.5",
+    "gatsby-plugin-twitter": "^2.2.4",
+    "gatsby-remark-autolink-headers": "^2.2.3",
+    "gatsby-remark-copy-linked-files": "^2.2.3",
+    "gatsby-remark-images": "^3.2.4",
     "gatsby-remark-prismjs": "^1.2.*",
-    "gatsby-remark-responsive-iframe": "^2.1.1",
-    "gatsby-source-filesystem": "^2.0.32",
+    "gatsby-remark-responsive-iframe": "^2.3.3",
+    "gatsby-source-filesystem": "^2.2.4",
     "gatsby-transformer-remark": "^1.7.*",
-    "gatsby-transformer-sharp": "^2.1.18",
-    "gh-pages": "^2.0.1",
-    "prettier": "^1.17.0",
-    "remark-cli": "^6.0.1",
-    "remark-preset-lint-recommended": "^3.0.2",
+    "gatsby-transformer-sharp": "^2.4.6",
+    "gh-pages": "^2.2.0",
+    "prettier": "^2.0.4",
+    "remark-cli": "^8.0.0",
+    "remark-preset-lint-recommended": "^4.0.0",
     "stylefmt": "^6.0.3",
-    "stylelint": "^10.0.1",
-    "stylelint-config-standard": "^18.3.0",
-    "write-good": "^1.0.1"
+    "stylelint": "^13.3.2",
+    "stylelint-config-standard": "^20.0.0",
+    "write-good": "^1.0.2"
   },
   "keywords": [
     "gatsby"
