@@ -4,27 +4,28 @@
     "email": "michelangelo.partipilo@gmail.com"
   },
   "dependencies": {
-    "cockpit-sdk": "^0.9.4",
+    "cockpit-sdk": "^0.15.1",
     "gatsby-node-helpers": "^0.3.0",
-    "graphql-type-json": "^0.2.1",
-    "html-to-react": "^1.3.3",
-    "pluralize": "^7.0.0",
-    "promise-throttle": "^1.0.0",
+    "graphql-type-json": "^0.3.1",
+    "html-to-react": "^1.4.2",
+    "pluralize": "^8.0.0",
+    "promise-throttle": "^1.0.1",
     "react-styling": "^1.6.4",
-    "sanitize-html": "^1.19.1",
+    "sanitize-html": "^1.23.0",
     "simple-rate-limiter": "^0.2.3",
     "valid-url": "^1.0.9"
   },
   "description": "Gatsby source plugin for Cockpit headless CMS API.",
   "devDependencies": {
-    "babel-eslint": "^10.0.1",
-    "eslint": "^5.8.0",
-    "eslint-config-fbjs": "^2.1.0",
-    "eslint-plugin-babel": "^5.2.1",
-    "eslint-plugin-flowtype": "^3.2.0",
-    "eslint-plugin-jsx-a11y": "^6.1.2",
-    "eslint-plugin-markdown": "^1.0.0-rc.0",
-    "eslint-plugin-react": "^7.11.1"
+    "babel-eslint": "^10.1.0",
+    "eslint": "^6.8.0",
+    "eslint-config-fbjs": "^3.1.1",
+    "eslint-config-prettier": "^6.10.1",
+    "eslint-plugin-babel": "^5.3.0",
+    "eslint-plugin-flowtype": "^4.7.0",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-markdown": "^1.0.2",
+    "eslint-plugin-react": "^7.19.0"
   },
   "keywords": [
     "gatsby",
