@@ -20,8 +20,10 @@ module.exports = class CockpitHelpers {
     this.reporter.info(`Specified Collections:`);
     // this.reporter.list(`collection names`, collections);
     // this.tick = this.reporter.progress(collections.length);
-    
-    return Promise.all(collections.map(name => this.getCollectionItems(name)));
+
+    return Promise.all(
+      collections.map((name) => this.getCollectionItems(name))
+    );
   }
 
   async getCollectionNames() {
@@ -30,36 +32,35 @@ module.exports = class CockpitHelpers {
 
     return explictlyDefinedCollections instanceof Array
       ? allCollections.filter(
-        name => explictlyDefinedCollections.indexOf(name) > -1
-      )
+          (name) => explictlyDefinedCollections.indexOf(name) > -1
+        )
       : allCollections;
   }
 
-  // get cockpit collection items by collection name
-  async getRegionItems(name) {
-    const values = await this.cockpit.regionData(name);
-    const template = await this.cockpit.regionGet(name);
-    return { data: {values, template}, name };
+  // get list of singletons
+  async getSingletonItems(name) {
+    const template = await this.cockpit.singletonGet(name);
+    return { data: { template }, name };
   }
 
-  // get all cockpit regions, together with their items
-  async getCockpitRegions() {
-    const regions = await this.getRegionNames();
-    return Promise.all(regions.map(name => {
-      var i = this.getRegionItems(name);
-      return i;
-    }));
-  }  
-
-  async getRegionNames() {
-    
-    const allRegions = await this.cockpit.regionList(); 
-    const explictlyDefinedRegions = this.config.regions;
-
-    return explictlyDefinedRegions instanceof Array
-      ? allRegions.filter(
-        name => explictlyDefinedRegions.indexOf(name) > -1
-      )
-      : allRegions;
+  // get all cockpit singletons, together with their items
+  async getCockpitSingletons() {
+    const items = await this.getSingletonNames();
+    return Promise.all(
+      items.map((name) => {
+        return this.getSingletonItems(name);
+      })
+    );
   }
-}
+
+  async getSingletonNames() {
+    const allSingletons = await this.cockpit.singletonList();
+    const explictlyDefinedSingletons = this.config.singletons;
+
+    return explictlyDefinedSingletons instanceof Array
+      ? allSingletons.filter(
+          (name) => explictlyDefinedSingletons.indexOf(name) > -1
+        )
+      : allSingletons;
+  }
+};
