@@ -1,5 +1,5 @@
 import React from "react";
-import Helmet from "react-helmet";
+import { Helmet } from "react-helmet";
 import _ from "lodash";
 
 import "../../scss/index.scss";
@@ -13,7 +13,7 @@ function capitalize(string) {
   return string.charAt(0).toUpperCase() + string.slice(1);
 }
 
-export default class MainLayout extends React.Component {
+class MainLayout extends React.Component {
   getInvertHeader() {
     const { location, invert } = this.props;
     if (invert) {
@@ -66,7 +66,7 @@ export default class MainLayout extends React.Component {
     const { children } = this.props;
     const invertFixedHeader = this.getInvertHeader();
     return (
-      <React.Fragment>
+      <>
         <Helmet>
           <title>{`${config.siteTitle} | ${this.getLocalTitle()}`}</title>
           <meta name="description" content={config.siteDescription} />
@@ -75,7 +75,9 @@ export default class MainLayout extends React.Component {
         <Header alwaysShow={invertFixedHeader} />
         {children}
         <Footer />
-      </React.Fragment>
+      </>
     );
   }
 }
+
+export default MainLayout;
