@@ -1,21 +1,17 @@
 import React, { Component } from "react";
 import {
   FacebookShareButton,
-  GooglePlusShareButton,
   LinkedinShareButton,
   TwitterShareButton,
   TelegramShareButton,
   RedditShareButton,
   FacebookShareCount,
-  GooglePlusShareCount,
-  LinkedinShareCount,
   RedditShareCount,
   FacebookIcon,
   TwitterIcon,
   TelegramIcon,
-  GooglePlusIcon,
   LinkedinIcon,
-  RedditIcon
+  RedditIcon,
 } from "react-share";
 
 import config from "../../../data/SiteConfig";
@@ -28,25 +24,19 @@ class SocialLinks extends Component {
     const url = config.siteUrl + config.pathPrefix + postPath;
 
     const iconSize = mobile ? 36 : 48;
-    const filter = count => (count > 0 ? count : "");
+    const filter = (count) => (count > 0 ? count : "");
 
     return (
       <div className="social-links">
         <RedditShareButton url={url} title={post.title}>
           <RedditIcon round size={iconSize} />
           <RedditShareCount url={url}>
-            {count => <div className="share-count">{filter(count)}</div>}
+            {(count) => <div className="share-count">{filter(count)}</div>}
           </RedditShareCount>
         </RedditShareButton>
         <TwitterShareButton url={url} title={post.title}>
           <TwitterIcon round size={iconSize} />
         </TwitterShareButton>
-        <GooglePlusShareButton url={url}>
-          <GooglePlusIcon round size={iconSize} />
-          <GooglePlusShareCount url={url}>
-            {count => <div className="share-count">{filter(count)}</div>}
-          </GooglePlusShareCount>
-        </GooglePlusShareButton>
         <FacebookShareButton
           url={url}
           title={post.title}
@@ -55,7 +45,7 @@ class SocialLinks extends Component {
         >
           <FacebookIcon round size={iconSize} />
           <FacebookShareCount url={url}>
-            {count => <div className="share-count">{filter(count)}</div>}
+            {(count) => <div className="share-count">{filter(count)}</div>}
           </FacebookShareCount>
         </FacebookShareButton>
         <LinkedinShareButton
@@ -64,9 +54,6 @@ class SocialLinks extends Component {
           description={postNode.excerpt}
         >
           <LinkedinIcon round size={iconSize} />
-          <LinkedinShareCount url={url}>
-            {count => <div className="share-count">{filter(count)}</div>}
-          </LinkedinShareCount>
         </LinkedinShareButton>
         <TelegramShareButton url={url}>
           <TelegramIcon round size={iconSize} />
