@@ -1,7 +1,7 @@
 import React from "react";
 import { graphql } from "gatsby";
 import Moment from "react-moment";
-import Helmet from "react-helmet";
+import { Helmet } from "react-helmet";
 import { Container, Row, Col } from "reactstrap";
 import Img from "gatsby-image";
 import Layout from "../components/Layout";
@@ -14,7 +14,7 @@ import config from "../../data/SiteConfig";
 
 export default class RecipeTemplate extends React.Component {
   render() {
-    const { data, pageContext } = this.props;
+    const { location, invert, data, pageContext } = this.props;
     const { slug } = pageContext;
     const { blog: postNode } = data;
     const { entry: post } = postNode;
@@ -24,7 +24,7 @@ export default class RecipeTemplate extends React.Component {
     const heroImage = postNode.entry.image.localFile.childImageSharp.fluid;
 
     return (
-      <Layout {...this.props}>
+      <Layout location={location} invert={invert}>
         <div className="blogPost">
           <Helmet>
             <title>{`${post.title} | ${config.siteTitle}`}</title>
@@ -44,7 +44,8 @@ export default class RecipeTemplate extends React.Component {
                   <div
                     dangerouslySetInnerHTML={{
                       __html:
-                        postNode.entry.content.markdown.childMarkdownRemark.html
+                        postNode.entry.content.markdown.childMarkdownRemark
+                          .html,
                     }}
                   />
                 </Col>
